preping evocache to save features in json to be sent to heroku db

+++ core/account.h
@@ -58,7 +58,7 @@ private:
 	void makeHashBundles() {
 		for (int i = 0; i < allTrans().count(); ++i) {
 			Transaction& t = allTrans().trans(i);
-			uint h = t.nameHash.hash;
+			qint64 h = t.nameHash.hash;
 			if (!m_hashBundles.contains(h))
 				m_hashBundles[h] = new TransactionBundle();
 			m_hashBundles[h]->append(&t);
